separate files, add gc

+++ src/constants.rs
@@ -1,21 +1,21 @@
 pub(crate) mod window_constants {
     pub(crate) const WINDOW_TITLE: &str = "Sand Tetris";
-    pub(crate) const WINDOW_WIDTH: u16 = 500;
-    pub(crate) const WINDOW_HEIGHT: u16 = 800;
+    pub(crate) const WINDOW_WIDTH: usize = 500;
+    pub(crate) const WINDOW_HEIGHT: usize = 800;
 }
 
 pub(crate) mod map_constants {
     use super::window_constants;
 
-    pub(crate) const GRAIN_SIDE_SIZE: u16 = 2;
-    pub(crate) const MAP_WIDTH: u16 = {
+    pub(crate) const GRAIN_SIDE_SIZE: usize = 2;
+    pub(crate) const MAP_WIDTH: usize = {
         assert!(
             window_constants::WINDOW_WIDTH % GRAIN_SIDE_SIZE == 0,
             "WINDOW_WIDTH is not divisible by GRAIN_SIDE_SIZE"
         );
         window_constants::WINDOW_WIDTH / GRAIN_SIDE_SIZE
     };
-    pub(crate) const MAP_HEIGHT: u16 = {
+    pub(crate) const MAP_HEIGHT: usize = {
         assert!(
             window_constants::WINDOW_HEIGHT % GRAIN_SIDE_SIZE == 0,
             "WINDOW_HEIGHT is not divisible by GRAIN_SIDE_SIZE"
